findBatchFiles added
This fixed all of the errors with the TP_FN files. The script now takes 5 minutes at the beginning to find all the files with _Batch as the suffix on the A: drive.

Also added more text to let the user know what is going on.

Also fixed issues with North Carolina TP_FN (spoiler alert - it has a space in its name). Couldn't fix the issues with NC FP, because there aren't any _Batch files in the entire NC gdb.