Refactored renaming_Batch_files.py
renaming_Batch_files.py is now more robust with error prevention and requires 1 input folder instead of 1 input folder and 1 input GDB.  It also has a better method of determining the new name for the copied file. The script is much faster to run and is fully fuctional a custom Python ArcGIS Tool.

In Automated_Review, regional_35_counties was added so whenever I want to run it for those counties I don't have to find the name of each of the clusters that contain counties from the 35 counties in the Regional Model.