added functionality to simulatedSampling
simulatedSampling function now deletes old iteration files instead of just replacing files on a 1-for-1 basis. This prevents the problem of running a batch with 10 iterations, then changing something and running it with 5 iterations, leaving 1-5 updated and 6-10 unupdated files. It is kinda gross to look at, but ugly is better than error-prone.

The readAdjNASS function within simSampling was also changed to fix the "DeKalb" problem. Now it just uses nameFormat and .title() on everthing to make sure we are comparing apples to apples.

Also, markCountyAsCompleted was added. This creates a CSV if it doesn't exist and then simply writes a line with the completed county. This is a bit rough, but it works. This will probably get updated further in the future, but as of now it is functional.