Minor refactoring and adjustment
Set it up to recieve adjFLAPS instead of adjNASS, though the row values will have to be changed in the readAdjFLAPS(...) function.

Refactored the collapsePoints(...) function so that it no longer creates an Integrate intermediate file, because there really was no point to it. This made things quicker and simpler.

The script now adds in another file to the neg_masks list for WA, OR, and CA states.

Silly error in markCountyAsCompleted(...) was fixed.