Mostly style and formatting updates
Most of this update consists of the start of the process of making this document mostly PEP 8 compliant (https://www.python.org/dev/peps/pep-0008/#comments). I have converted notations at the beginning of functions to docstrings(https://www.python.org/dev/peps/pep-0257/), as well as fixing some indentation issues and changing annotations to be mostly PEP 8 compliant.

I still need to do the naming conventions, but this script is a lot more readable and consistant than before.

Pretty much the only functional change is that now LAR thresholds can be None, instead of just having an arbitrarily high number.  Also fixed a lot of if var == None errors (hint: it should be if var is None).